arabicmodel/requirements.txt
openai-whisper
python-pptx
requests
flask
torch